refactor: decouple productivity logic into useProductivityData hook
Extracted data formatting, task prioritization, and daily plan generation from ProductivityScreen into a dedicated custom hook to adhere to SOLID principles and resolve the "God Class" violation.
